Adicionado validação para impedir que um nível seja removido quando houver um (ou mais) desenvolvedor(es) associado a este

+++ dev-manager-api/src/nivel/nivel.controller.ts
@@ -4,13 +4,17 @@ import { CreateNivelDto } from './dto/create-nivel.dto';
 import { UpdateNivelDto } from './dto/update-nivel.dto';
 import { validate } from 'class-validator';
 import { Response } from 'express';
+import { DesenvolvedorService } from 'src/desenvolvedor/desenvolvedor.service';
 
 @Controller('nivel')
 export class NivelController {
-  constructor(private readonly nivelService: NivelService) {}
+  constructor(
+    private readonly nivelService: NivelService,
+    private readonly devService: DesenvolvedorService
+  ) { }
 
   @Post()
-  async create(@Body() createNivelDto: CreateNivelDto, @Res() res: Response): Promise<any>  {
+  async create(@Body() createNivelDto: CreateNivelDto, @Res() res: Response): Promise<any> {
     try {
       // const error = await validate(createNivelDto);
       const error = await validate(this.nivelService.nivelFromDTO(createNivelDto));
@@ -36,15 +40,15 @@ export class NivelController {
   @Put(':id')
   async update(@Param('id') id: string, @Body() updateNivelDto: UpdateNivelDto, @Res() res: Response): Promise<any> {
     const nivel = await this.nivelService.findOne(+id);
-    if (nivel){
-      try{
+    if (nivel) {
+      try {
         const nivel = await this.nivelService.update(+id, updateNivelDto);
         return res.status(HttpStatus.OK).json(nivel);
       } catch (error) {
         return res.status(HttpStatus.BAD_REQUEST).send(error);
       }
     } else {
-      return res.status(HttpStatus.NOT_FOUND).send("Nivel com id " + +id + " não encontrado!");  
+      return res.status(HttpStatus.NOT_FOUND).send("Nivel com id " + +id + " não encontrado!");
     }
   }
 
@@ -56,15 +60,20 @@ export class NivelController {
   @Delete(':id')
   async remove(@Param('id') id: string, @Res() res: Response) {
     const nivel = await this.nivelService.findOne(+id);
-    if (nivel){    
-      try {
-        await this.nivelService.remove(+id);
-        return res.status(HttpStatus.NO_CONTENT).send();
-      } catch (error) {
-        return res.status(HttpStatus.BAD_REQUEST).json(error);
+    if (nivel) {
+      const countDevs = await this.devService.count(+id);
+      if (countDevs > 0) {
+        return res.status(HttpStatus.NOT_IMPLEMENTED).send("Nivel com id " + +id + " possui desenvolvedores associados!");
+      } else {
+        try {
+          await this.nivelService.remove(+id);
+          return res.status(HttpStatus.NO_CONTENT).send();
+        } catch (error) {
+          return res.status(HttpStatus.BAD_REQUEST).json(error);
+        }
       }
     } else {
-      return res.status(HttpStatus.NOT_FOUND).send("Nivel com id " + +id + " não encontrado!");  
+      return res.status(HttpStatus.NOT_FOUND).send("Nivel com id " + +id + " não encontrado!");
     }
   }
 }
